% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{as.matrix.Dataset}
\alias{as.matrix.Dataset}
\title{Dataset in Matrix and data.frame}
\usage{
\method{as.matrix}{Dataset}(Dataset)
}
\arguments{
\item{Dataset:}{Element of class Dataset}
}
\description{
Dataset in Matrix and data.frame
}
\examples{
. 
as.matrix(Dataset(cars, target = "dist"))
as.dataframe(Dataset(cars, target = "dist"))
}
