% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_utils.R
\name{checkHyperparameter}
\alias{checkHyperparameter}
\title{check if Hyperparameters are valid for given hyperparameter space}
\usage{
checkHyperparameter(hps, hpx)
}
\arguments{
\item{hps}{}

\item{hpx}{}
}
\description{
check if Hyperparameters are valid for given hyperparameter space
}
