% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{Dataset}
\alias{Dataset}
\title{Dataset Config}
\usage{
Dataset(
  data,
  target,
  type = character(0),
  name = deparse(substitute(data), 20)[[1]]
)
}
\arguments{
\item{data:}{Dataframe for target and features}

\item{target:}{Character, which represent the target name}

\item{type:}{"regression", "classification"}

\item{name:}{Character}
}
\description{
How is the dataset structured and how can it be manipulated
}
\examples{
. 
Dataset(cars, target = "dist", type = "regression")
}
