% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_utils.R
\name{hp}
\alias{hp}
\title{Constructor for class 'hp'}
\usage{
hp(...)
}
\arguments{
\item{...}{}
}
\description{
Constructor for class 'hp'
}
\examples{

hpx <- hp(p_dbl(0, 1), p_int(1, 10), p_fct(letters))
hpx

#>    name type                range
#> 1:    x  dbl               0, 1
#> 2:    y  int              1, 10
#> 3:    z  fct "a", "b", "c", ...
}
