% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{HyperparameterSpace}
\alias{HyperparameterSpace}
\alias{hp}
\title{Combine Hyperparameter objetcs into a Hyperparameter Space}
\usage{
HyperparameterSpace(...)

hp(...)
}
\arguments{
\item{...}{List of named or unnamed Hyperparameters}
}
\description{
If there are unnamed Hyperparameters given, we start naming them from x, y, z, a, ...
Unique names are ensured.
}
\examples{
HyperparameterSpace(p_int(0, 9), p_dbl(lower = 0))
hp(p_fct(category = c("a", "b", "c")), alpha = p_dbl(0, 1))
}
