% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inducer.R
\name{InducerRandomForest}
\alias{InducerRandomForest}
\title{Inducer for Random Forest}
\usage{
InducerRandomForest(
  .data,
  min.node.size = NULL,
  max.depth = NULL,
  num.trees = 500,
  mtry = NULL,
  verbose = 0,
  replace = TRUE
)
}
\arguments{
\item{.data}{dataset object of class 'Dataset'}

\item{min.node.size}{Minimal node size. Default 1 for classification, 5 for regression, 3 for survival and 10 for probability}

\item{max.depth}{Maximal tree depth. NULL or 0 corresponds to unlimited depth, 1 to tree stumps}

\item{num.trees}{Number of trees. Default 500 trees}

\item{mtry}{Number of variables to potentially split in each node. Default rounded down square root of the number variables}

\item{verbose}{If 0, no informations will be printed. If 1, it will print informations about performances. If 2,additional information will be printed.}

\item{replace}{Sample with replacement}
}
\description{
Inducer for Random Forest
}
\examples{
iris.data <- Dataset(iris, "Species", type = "classification")
rforest <- InducerRandomForest(num.trees = 400)
mod <- rforest(cars.data)

mod <- InducerRandomForest(cars.data, num.trees = 400)
}
