% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{length.SplitCVInstance}
\alias{length.SplitCVInstance}
\title{Length method for 'SplitCVInstance' class}
\usage{
\method{length}{SplitCVInstance}(x)
}
\arguments{
\item{x}{An object of class 'SplitCVInstance'}
}
\description{
Length method for 'SplitCVInstance' class
}
