% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{metainfo}
\alias{metainfo}
\title{Dataset Metainfo}
\usage{
metainfo(data)
}
\arguments{
\item{data:}{Element of class Dataset}
}
\description{
Dataset Metainfo
}
\examples{
. 
metainfo(Dataset(cars, target = "dist"))

$feature
$feature[[1]]
[1] "speed"   "numeric"


$target
[1] "dist"    "numeric"

$nrow
[1] 50

$type
character(0)

$missings
[1] FALSE

attr(,"class")
[1] "DatasetInfo"
}
