% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{nrow.Dataset}
\alias{nrow.Dataset}
\title{The Number of Rows of a \code{Dataset} object}
\usage{
\method{nrow}{Dataset}(x)
}
\arguments{
\item{x}{The Dataset}
}
\description{
Uses \code{base::nrow} under the hood
}
