% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{nrow.default}
\alias{nrow.default}
\title{The Number of Rows of an Array}
\usage{
\method{nrow}{default}(x)
}
\arguments{
\item{x}{The array/object to return rows for}
}
\description{
The Number of Rows of an Array
}
