% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemaker.R
\name{p_fct}
\alias{p_fct}
\title{make hyperparameter space}
\usage{
p_fct(levels)
}
\arguments{
\item{factors}{vector of factors}
}
\description{
This function returns the type, and start and end of a vector of factors
}
\examples{
x <- p_fct(c("a", "b"))

x

#> $type
#> [1] "fct"

#> $factors
#> [1] "a" "b"
}
