% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{predict.ModelLinearModel}
\alias{predict.ModelLinearModel}
\title{Predict with a Linear Model}
\usage{
\method{predict}{ModelLinearModel}(object, newdata, ...)
}
\arguments{
\item{object}{The model object to predict with}

\item{newdata}{New data to make predictions on}

\item{...}{Further arguments passed to \code{predict}}
}
\description{
Predict with a Linear Model
}
