% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{print.Dataset}
\alias{print.Dataset}
\title{Dataset Print}
\usage{
\method{print}{Dataset}(object)
}
\arguments{
\item{object:}{Element of class Dataset}
}
\description{
Dataset Print
}
\examples{
. 
print(Dataset(cars, target = "dist", type = "regression"))
Dataset "cars", predicting "dist" (regression)

   speed dist
1      4    2
2      4   10
3      7    4
}
