% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{print.Split}
\alias{print.Split}
\title{Print method for 'Split' class}
\usage{
\method{print}{Split}(x, ...)
}
\arguments{
\item{x}{An object of class 'Split'}

\item{...}{Other arguments passed to 'print'}
}
\description{
Print method for 'Split' class
}
