% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{print.SplitInstance}
\alias{print.SplitInstance}
\title{Print method for 'SplitInstance' class}
\usage{
\method{print}{SplitInstance}(x, ...)
}
\arguments{
\item{x}{An object of class 'SplitInstance'}

\item{...}{Other arguments passed to 'print'}
}
\description{
Print method for 'SplitInstance' class
}
