% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample}
\alias{resample}
\title{Resampler function}
\usage{
resample(dataset, inducer, splitClass)
}
\arguments{
\item{dataset}{The Dataset object to resample}

\item{inducer}{The Inducer Model to use for prediction}

\item{splitClass}{The Split class to use}
}
\description{
Resampler function
}
