% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{[.Dataset}
\alias{[.Dataset}
\title{Dataset column and row selection}
\usage{
\method{[}{Dataset}(obj, i, j = names(obj$data))
}
\arguments{
\item{obj:}{Element of class Dataset}

\item{i:}{row}

\item{j:}{column}
}
\description{
Dataset column and row selection
}
\examples{
. 
Dataset(cars, target = "dist")[c(1,2),]

speed dist
1     4    2
2     4   10
}
