% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{[[.SplitCVInstance}
\alias{[[.SplitCVInstance}
\title{Access fold i of 'SplitCVInstance' class}
\usage{
\method{[[}{SplitCVInstance}(x, i)
}
\arguments{
\item{x}{An object of class 'SplitCVInstance'}

\item{i}{The index of the fold to access}
}
\description{
Access fold i of 'SplitCVInstance' class
}
