% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemaker.R
\docType{data}
\name{x}
\alias{x}
\title{make hyperparameter space}
\format{
An object of class \code{list} of length 3.
}
\usage{
x
}
\arguments{
\item{range1}{start (left) value}

\item{range2}{end (right) value}

\item{factors}{vector of factors}
}
\description{
These functions returns the type, and start and end of a range for dbl and int, or a vector of factors for fct
}
\examples{
x <- p_dbl(1,3)

x

#> $type
#> [1] "dbl"

#> $start
#> [1] 1

#> $end
#> [1] 3
}
\keyword{datasets}
