% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{checkHyperparameter}
\alias{checkHyperparameter}
\title{Check if list values lie in a given HyperparameterSpace}
\usage{
checkHyperparameter(hp, hp.space)
}
\arguments{
\item{hp}{List of named Hyperparameter values}

\item{hp.space}{HyperparameterSpace used as reference}
}
\value{
List with names of \code{hp} and logical values indicating if the value lies in the given HyperparameterSpace
}
\description{
This checks if a named lists values lie in a given HyperparameterSpace.
}
\examples{
hp.space <- hp(x = p_int(0, 9), y = p_dbl(lower = 0))
checkHyperparameter(list(x = 5, y = 2.5), hp.space)
}
