% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{Dataset}
\alias{Dataset}
\title{Creation class: "Dataset"}
\usage{
Dataset(
  data,
  target,
  type = c("regression", "classification"),
  name = deparse(substitute(data), 20)
)
}
\arguments{
\item{data}{Data, either a data.table or data.frame}

\item{target}{Target column}

\item{type}{Type of the task. One of c("regression", "classification")}

\item{name}{Name of the dataset}
}
\description{
Creation class: "Dataset"
}
