% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{p_dbl}
\alias{p_dbl}
\title{Double Hyperparameter}
\usage{
p_dbl(lower = -Inf, upper = Inf)
}
\arguments{
\item{lower}{Lower bound of the hyperparameter. Defaults to -Inf.}

\item{upper}{Upper bound of the hyperparameter. Defaults to Inf.}
}
\description{
Generate a Hyperparameter of type \code{Double}
}
