% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{p_int}
\alias{p_int}
\title{Integer Hyperparameter}
\usage{
p_int(lower = -Inf, upper = Inf)
}
\arguments{
\item{lower}{Lower bound of the hyperparameter. Defaults to -Inf.}

\item{upper}{Upper bound of the hyperparameter. Defaults to Inf.}
}
\description{
Generate a Hyperparameter of type \code{Integer}
}
